Move target.py and descriptor.py into the core module.
   - Update all impacted imports
