+++ numba_dpex/core/descriptor.py
@@ -14,12 +14,12 @@ class DpexTarget(TargetDescriptor):
 
     @utils.cached_property
     def _toplevel_target_context(self):
-        # Lazily-initialized top-level target context, for all threads
+        """Lazily-initialized top-level target context, for all threads."""
         return DpexTargetContext(self.typing_context, self._target_name)
 
     @utils.cached_property
     def _toplevel_typing_context(self):
-        # Lazily-initialized top-level typing context, for all threads
+        """Lazily-initialized top-level typing context, for all threads."""
         return DpexTypingContext()
 
     @property
