+++ numba_dpex/spirv_generator.py
@@ -10,7 +10,7 @@ import tempfile
 from subprocess import CalledProcessError, check_call
 
 from numba_dpex import config
-from numba_dpex.target import LINK_ATOMIC, LLVM_SPIRV_ARGS
+from numba_dpex.core.target import LINK_ATOMIC, LLVM_SPIRV_ARGS
 
 
 def _raise_bad_env_path(msg, path, extra=None):
